+++ CHANGELOG.md
@@ -3,5 +3,5 @@
 **New Features**
 
 * First preview release of the Azure SDK Document Classifier Model.
-* Exposes preliminary API via `AzureSDKTrack1Classifier.is_t1(str) -> bool`.
-    - Exposes verbose API via `AzureSDKTrack1Classifier.is_t1_verbose(str) -> dict` to obtain classification metadata.
+* Exposes preliminary API via `AzureSDKTrackClassifier.is_t1(str) -> bool`.
+    - Exposes verbose API via `AzureSDKTrackClassifier.is_t1_verbose(str) -> dict` to obtain classification metadata.